Update for 2401 and consolidate guest tool checks

+++ src/ubuntu/chef/cookbooks/gusztavvargadr_packer_ubuntu/recipes/initialize.rb
@@ -1,42 +1,48 @@
-if guest?
-  apt_update '' do
-    action :update
-  end
+apt_update '' do
+  action :update
 end
 
 if vbox?
-  apt_package [ 'build-essential', 'cryptsetup', 'libssl-dev', 'libreadline-dev', 'zlib1g-dev', 'linux-source', 'dkms', 'linux-headers-generic' ] do
-    action :upgrade
-    notifies :run, 'bash[guest-additions]', :immediately
-    notifies :request_reboot, 'reboot[gusztavvargadr_packer_ubuntu]', :immediately
-  end
+  vbox_version = shell_out('VBoxControl -v').stdout.strip
 
-  bash 'guest-additions' do
-    code <<-EOH
-      VER="`cat /home/vagrant/.vbox_version`";
-      ISO="VBoxGuestAdditions_$VER.iso";
-      wget http://download.virtualbox.org/virtualbox/$VER/$ISO
-      mkdir -p /tmp/vbox;
-      mount -o loop $ISO /tmp/vbox;
-      sh /tmp/vbox/VBoxLinuxAdditions.run \
-          || echo "VBoxLinuxAdditions.run exited $? and is suppressed." \
-              "For more read https://www.virtualbox.org/ticket/12479";
-      umount /tmp/vbox;
-      rm -rf /tmp/vbox;
-      rm -f *.iso;
+  unless vbox_version.include?('6.')
+    apt_package [ 'build-essential', 'cryptsetup', 'libssl-dev', 'libreadline-dev', 'zlib1g-dev', 'linux-source', 'dkms', 'linux-headers-generic' ] do
+      action :upgrade
+      notifies :run, 'bash[guest-additions]', :immediately
+      notifies :request_reboot, 'reboot[gusztavvargadr_packer_ubuntu]', :immediately
+    end
+
+    bash 'guest-additions' do
+      code <<-EOH
+        VER="`cat /home/vagrant/.vbox_version`";
+        ISO="VBoxGuestAdditions_$VER.iso";
+        wget http://download.virtualbox.org/virtualbox/$VER/$ISO
+        mkdir -p /tmp/vbox;
+        mount -o loop $ISO /tmp/vbox;
+        sh /tmp/vbox/VBoxLinuxAdditions.run \
+            || echo "VBoxLinuxAdditions.run exited $? and is suppressed." \
+                "For more read https://www.virtualbox.org/ticket/12479";
+        umount /tmp/vbox;
+        rm -rf /tmp/vbox;
+        rm -f *.iso;
 EOH
-    action :nothing
+      action :nothing
+    end
   end
 end
 
 if vmware?
-  apt_package [ 'open-vm-tools', 'open-vm-tools-desktop' ] do
-    action :upgrade
-    notifies :request_reboot, 'reboot[gusztavvargadr_packer_ubuntu]', :immediately
+  vmware_version = shell_out('vmware-toolbox-cmd -v').stdout.strip
+
+  unless vmware_version.include?('12.')
+    apt_package [ 'open-vm-tools', 'open-vm-tools-desktop' ] do
+      action :upgrade
+      notifies :request_reboot, 'reboot[gusztavvargadr_packer_ubuntu]', :immediately
+    end
   end
 end
 
-if guest? && !vbox? && !vmware?
+if !vbox? && !vmware?
   apt_package [ 'linux-image-virtual', 'linux-tools-virtual', 'linux-cloud-tools-virtual' ] do
     action :upgrade
     notifies :request_reboot, 'reboot[gusztavvargadr_packer_ubuntu]', :immediately
